+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @globalfishingwatchapp/fishing-map
 
+## 1.1.12
+
+### Patch Changes
+
+- Updated dependencies [720f160a]
+  - @globalfishingwatch/layer-composer@8.10.4
+
+## 1.1.11
+
+### Patch Changes
+
+- Updated dependencies [828f0c4b]
+  - @globalfishingwatch/layer-composer@8.10.3
+
 ## 1.1.10
 
 ### Patch Changes