+++ applications/vessel-history/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @globalfishingwatchapp/vessel-history
 
+## 0.0.76
+
+### Patch Changes
+
+- Updated dependencies [720f160a]
+  - @globalfishingwatch/layer-composer@8.10.4
+
+
 ## 0.0.75
 
 ### Patch Changes
@@ -7,6 +15,8 @@
 - Include resources slice, selectors and helpers to consume API
 - Updated dependencies [undefined]
   - @globalfishingwatch/dataviews-client@11.8.1
+- Updated dependencies [828f0c4b]
+  - @globalfishingwatch/layer-composer@8.10.3
 
 ## 0.0.74